Refactor AssignServerRoles
Move all components specific to this page into pages/AssignServerRoles.
Combine the .less files of each of these subpages into a single .less
file for AssignServerRoles.
