+++ src/pages/AssignServerRoles.js
@@ -7,15 +7,15 @@ import { getAppConfig } from '../utils/ConfigHelper.js';
 import { ActionButton, LoadFileButton } from '../components/Buttons.js';
 import { IpV4AddressValidator, MacAddressValidator } from '../utils/InputValidators.js';
 import { SearchBar, ServerRolesAccordion, ServerInputLine, ServerDropdownLine, isRoleAssignmentValid }
-  from '../components/ServerUtils.js';
+  from './AssignServerRoles/ServerUtils.js';
 import { BaseInputModal, ConfirmModal } from '../components/Modals.js';
 import BaseWizardPage from './BaseWizardPage.js';
-import ConnectionCredsInfo from '../components/ConnectionCredsInfo';
+import ConnectionCredsInfo from './AssignServerRoles/ConnectionCredsInfo';
 import { ErrorMessage } from '../components/Messages.js';
 import { LoadingMask } from '../components/LoadingMask.js';
-import ServerTable from '../components/ServerTable.js';
-import EditServerDetails from '../components/EditServerDetails.js';
-import ViewServerDetails from '../components/ViewServerDetails';
+import ServerTable from './AssignServerRoles/ServerTable.js';
+import EditServerDetails from './AssignServerRoles/EditServerDetails.js';
+import ViewServerDetails from './AssignServerRoles/ViewServerDetails';
 import { importCSV } from '../utils/CsvImporter.js';
 import { alphabetically } from '../utils/Sort.js';
 import { fromJS } from 'immutable';
