+++ src/pages/AssignServerRoles/BaseServerDetails.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Tabs, Tab } from 'react-bootstrap';
-import { translate } from '../localization/localize.js';
+import { translate } from '../../localization/localize.js';
 import ModelServerDetails from './ModelServerDetails.js';
 
 const DETAILS_FROM_SERVER_TAB = 1;
