+++ src/pages/AssignServerRoles/EditServerDetails.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
-import { translate } from '../localization/localize.js';
-import { ActionButton } from './Buttons.js';
-import { ServerInputLine, ServerDropdownLine} from '../components/ServerUtils.js';
+import { translate } from '../../localization/localize.js';
+import { ActionButton } from '../../components/Buttons.js';
+import { ServerInputLine, ServerDropdownLine} from './ServerUtils.js';
 import {
   IpV4AddressValidator, MacAddressValidator
-} from '../utils/InputValidators.js';
+} from '../../utils/InputValidators.js';
 
 const UNKNOWN = -1;
 const VALID = 1;
