+++ src/pages/AssignServerRoles/OvServerDetails.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { translate } from '../localization/localize.js';
+import { translate } from '../../localization/localize.js';
 import BaseServerDetails from './BaseServerDetails.js';
 
 class OvServerDetails extends BaseServerDetails {
