--- src/styles/components/connetioncredsinfo.less
@@ -1,24 +1,0 @@
-@import "./../palette.less";
-
-.connection-creds-info {
-  min-height: 38em;
-
-  .creds-category {
-    margin: 0.5em 0.5em 1.5em 0;
-  }
-
-  .secured {
-    margin: 0.5em 0.5em 1.5em 14.5em;
-  }
-
-  .input-button-container {
-    position: absolute;
-    margin-top: 1em;
-    right: 2em;
-    bottom: 1em;
-  }
-}
-
-.creds-dialog {
-  min-width: 55em;
-}
