--- src/styles/components/editserverdetails.less
@@ -1,14 +1,0 @@
-@import "./../palette.less";
-
-.edit-server-details {
-  min-height: 37em;
-  .input-button-container {
-    position: absolute;
-    margin-top: 1em;
-    right: 2em;
-    bottom: 1em;
-  }
-}
-.edit-details-dialog {
-  min-width: 55em;
-}
