--- src/styles/components/serverutils.less
@@ -1,150 +1,0 @@
-@import "./../palette.less";
-
-.roles-accordion {
-  height: 100%;
-
-  //overwrite default classes of collapsible
-  .Collapsible__contentInner {
-    border-top: 0;
-    background: white;
-    height: 9em;
-  }
-
-  .Collapsible__contentOuter {
-    overflow: auto !important;
-  }
-
-  .Collapsible__trigger {
-    background: linear-gradient(to bottom, @table-header-gradient-color 0%, @table-header-color 100%);
-    border-top: 1px solid @box-border-color;
-    border-bottom: 1px solid @box-border-color;
-    display: block;
-    font-weight: 400;
-    text-decoration: none;
-    color: @detailcolor;
-    position: relative;
-    padding: 10px;
-
-    &:after {
-      position: absolute;
-      right: 10px;
-      top: 10px;
-      display: block;
-      transition: transform 300ms;
-    }
-
-    &.is-open {
-      &:after {
-        transform: rotateZ(180deg);
-      }
-    }
-
-    &.has-error {
-      color: @errorcolor;
-    }
-  }
-}
-
-.search-container {
-  display: inline-block;
-  margin-bottom: 1em;
-  position:relative;
-  width: 100%;
-}
-
-.search-bar input {
-  width: 100%;
-  font-size: 13px;
-  padding: 0.5em;
-}
-
-.search-icon {
-  position: absolute;
-  top: 0.7em;
-  right: 0.5em;
-}
-
-.server-details-container {
-  .detail-line {
-    width: 100%;
-    margin-bottom: 0.7em;
-    display: flex;
-    .detail-heading {
-      width: 35%;
-      min-width: 10em;
-      display: inline-block;
-      text-align: right;
-      padding-right: 2em;
-      &.bold {
-        font-weight: 600;
-      }
-    }
-    .info-body {
-      width: 80%;
-      min-width: 30em;
-      display: inline-block;
-      position: relative;
-      top: 0px;
-      margin-bottom: 0.6em;
-      &.readonly {
-        min-height: 1.5em;
-      }
-    }
-    .input-body {
-      width: 80%;
-      min-width: 30em;
-      display: inline-block;
-      position: relative;
-      top: -4px;
-    }
-  }
-
-  .server-detail-select select {
-    padding: 4px;
-    width: 80%;
-  }
-
-  &.viewonly {
-    margin-top: 4em;
-  }
-}
-
-.message-line {
-  line-height: 2em;
-  margin: 3em 0 1em 1em;
-}
-
-.server-input {
-  .error-message {
-    color:@errorcolor;
-    display: block;
-  }
-}
-
-.server-input input {
-  padding: 4px;
-  width: 80%;
-}
-
-.sm-badge:before {
-  content: ' \25CF';
-  font-size: 1em;
-  margin-right: 0.5em;
-  color: @preferredcolor;
-}
-
-.ov-badge:before {
-  content: ' \25CF';
-  font-size: 1em;
-  margin-right: 0.5em;
-  color: @preferredcolor-hpe;
-}
-
-.material-icons.password-icon {
-  font-size: 16px;
-  margin-left: 5px;
-  cursor: pointer;
-  position: relative;
-  left: -27px;
-  top: 2px;
-}
