--- src/styles/components/svrroleassigner.less
@@ -1,84 +1,0 @@
-@import "./../palette.less";
-
-.assign-server-role {
-  .body-container {
-    margin: 2em;
-  }
-
-  .server-container{
-    display: inline-block;
-    min-height: 300px;
-    width: 49%;
-    vertical-align: top;
-  }
-
-  .right-col {
-    float: right;
-  }
-
-  .action-line {
-    display: table;
-    .action-item-left {
-      display: table-cell;
-      width: 100%;
-    }
-    .action-item-right {
-      display: inline-flex;
-      margin-left: 0.5em;
-    }
-  }
-
-  .search-container {
-    margin-bottom: 1em;
-    position:relative;
-  }
-
-  .table-header-wrapper {
-    display: flex;
-  }
-
-  .search-bar input {
-    width: 100%;
-    font-size: 13px;
-    padding: 0.5em;
-  }
-
-  .search-icon {
-    position: absolute;
-    top: 0.7em;
-    right: 0.5em;
-  }
-
-  .server-table-container {
-    position: relative;
-    width: 100%;
-    min-height: 30em;
-    height: 40em;
-    background: white;
-    margin-bottom: 1em;
-    .tab-pane {
-      margin-top: 0.5em;
-      max-height: 36em;
-      overflow: auto;
-    }
-    &.server-dropzone{
-      margin: 2px;
-      outline-offset: 1px;
-    }
-  }
-
-  .centered {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-  }
-
-  .role-accordion-container {
-    height: 43.5em;
-  }
-}
-
-.manual-discover-modal {
-  min-width: 55em;
-}
