--- src/styles/components/viewserverdetails.less
@@ -1,59 +1,0 @@
-@import "./../palette.less";
-
-.view-server-details {
-  min-height: 60em;
-  overflow-y: auto;
-  .button-container {
-    position: absolute;
-    margin-top: 1em;
-    right: 2em;
-    bottom: 1em;
-  }
-
-  .panel-default {
-    border-color: @box-border-color;
-  }
-
-  .panel {
-    background-color: white;
-    border: 1px solid transparent;
-  }
-
-
-  .panel-default > .panel-heading {
-    color: @page-font-color;
-    background: linear-gradient(to bottom, @table-header-gradient-color 0%, @table-header-color 100%);
-    border-color: @box-border-color;
-    line-height: 2.8em;
-    padding: 0.5em 0 0.5em 1em;
-  }
-
-  .panel-body {
-    padding: 3em;
-  }
-
-  .table {
-    border-collapse: separate;
-    tbody > tr > th {
-      border-top: none;
-    }
-    thead > tr > th {
-      background: white;
-    }
-  }
-
-  &.shorter-height {
-    min-height: 35em;
-  }
-}
-
-.view-details-dialog {
-  min-width: 50em;
-  &.more-width {
-    min-width: 80em;
-  }
-  .tab-pane {
-    max-height: 68vh;
-  }
-}
-
