+++ src/styles/deployer.less
@@ -11,19 +11,15 @@
 @import "components/ymlvalidator.less";
 @import "components/playbookprogress.less";
 @import "components/buttons.less";
-@import "components/svrroleassigner.less";
 @import "components/serverprovision.less";
 @import "components/transfertable.less";
-@import "components/editserverdetails.less";
 @import "components/contextmenu.less";
 @import "components/modals.less";
 @import "components/logviewer.less";
-@import "components/serverutils.less";
 @import "components/collapsibletable.less";
-@import "components/connetioncredsinfo.less";
 @import "components/messages.less";
 @import "components/loadingmask.less";
-@import "components/viewserverdetails";
 
 // Pages
+@import "pages/AssignServerRoles.less";
 @import "pages/ServerRoleSummary.less";
